Unify stats views + deterministic column_id resolution (#90)
This change refactors connector statistics production to use a shared
ConnectorStatsViewBuilder and introduces deterministic column_id
computation via ColumnIdAlgorithm (Iceberg: CID_FIELD_ID,
Delta: CID_PATH_ORDINAL).

+++ service/src/main/java/ai/floedb/floecat/service/repo/model/ColumnStatsKey.java
@@ -17,5 +17,5 @@
 package ai.floedb.floecat.service.repo.model;
 
 public record ColumnStatsKey(
-    String accountId, String tableId, long snapshotId, int columnId, String sha256)
+    String accountId, String tableId, long snapshotId, long columnId, String sha256)
     implements ResourceKey {}
